Remove compiler warnings Add missing DBUG_RETURN Fixed stack overflow in NdbBlob (found by ndb_gis.test) Fixed access to freed memory in ndb_cluster_real_free_share()

+++ storage/ndb/src/ndbapi/NdbScanOperation.cpp
@@ -169,7 +169,7 @@ NdbScanOperation::readTuples(NdbScanOperation::LockMode lm,
   { char* p = getenv("NDB_USE_TUPSCAN");
     if (p != 0) {
       unsigned n = atoi(p); // 0-10
-      if (::time(0) % 10 < n) tupScan = true;
+      if ((unsigned int) (::time(0) % 10) < n) tupScan = true;
     }
   }
 #endif
